Move Scope to its own file
